requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
nltk==3.8.1
scikit-learn==1.3.2
textblob>=0.15.3
vaderSentiment>=3.3.2
plotly>=5.3.0 